refactor: hooks file naming
